+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## features:
 
+* new `geom_pie()` function to draw pie charts
+
 * new `ggwrap()` function to wrap arbitrary graphics to the layout
 
 * new `fortify_stack()` function to convert any objects for plot with [ggstack()]